[5.10] 索引与计数统一 int64：nonzero / randperm / topk / 池化 indices，删掉 repeat_interleave 的 2^31 断言
torch 给每一个索引都用 int64，理由不是形式上的：索引就是缓冲区里的偏移量，
int32 到 2^31 个元素就叫不出名字了。jittor 这几处返回 int32：

- `jt.where(cond)` / `nonzero`（`where_op.h` 的默认 dtype）
- `randperm`
- `topk` 的索引（空输入那条分支自己造数组，非空那条带着 argsort 的输出）
- `MaxPool2d/3d`、`AdaptiveMaxPool2d/3d` 的 `return_indices`

而且 dtype 不只是标签：jittor 按字节宽度提升，所以 `index * stride` 停在索引自己的
dtype 里。四个元素、步长 1e9 就够了——int32 给出 -1294967296，int64 给出 3000000000。
测试里保留了这一条，因为它是「dtype 就是缺陷本身」的最短证明。

顺带修的三处（都是改 dtype 之后才暴露出来的、原本被 int32 掩盖的东西）：

1. `cub_where_op.cc` 里所有计数原本是 `int`/`index_t`。To 一旦不是 int32，
   `std::min(1024, num_nonzeros_h)` 连编译都过不去；`free(num_nonzeros, sizeof(int), ...)`
   与上面的 `alloc(sizeof(To), ...)` 也只在 To 是 int32 时数字相同。计数改 int64，
   grid 改成按 65535 封顶（kernel 本来就是 grid-stride 循环）。
2. 池化的索引编码 `p * W + q`（3D 是 `(p*H+q)*W+r`）是**一个 (n,c) 平面内的展平偏移**，
   原本在 `int` 里算，一个平面到 2^31 个元素就溢出——恰恰是索引最要紧的时候。改 int64。
3. `repeat_interleave` 的 CUDA 快路径把 repeats 转成 int32、在 int32 里做前缀和、
   用 `int out_row` 索引输出，然后用
   `assert output_size <= 2147483647` 把唯一做不了的情形挡在门外。这个断言是诚实的，
   所以「删断言」必须先把这条路径改成 64 位。取证（`tests/ops/test_index_dtype.py`
   里标了 `pytest.mark.manual` 的探针，24G 卡上 6.5 秒）：把算术改回 32 位、只删断言，
   2^31 之后的行全部返回第一行的值（`out_row` 变负，二分查找恒选 row 0）——
   也就是说断言原来挡住的是一个静默错答案，不是保守。

验证：`tests/ops/test_index_dtype.py`，CPU 与 CUDA 各一份，修前 18 个用例失败、修后全过。
期望 dtype 是在独立解释器里从 torch 2.12.1 读出来后固化成常量的（jittor 环境里的
torch 是 shim，不能当基准）。另外覆盖 MaxUnpool2d 的往返，确认 int64 索引仍能进 reindex 表达式。

范围之外，记录在此不一起改：
- `jt.argsort`、`jt.argmax`/`argmin`、`jt.arange` 仍返回 int32（torch 是 int64）。
  前两个是 C++ 算子默认值且底下挂着 CUB 路径，arange 的影响面覆盖整个库。
  topk 因此是 cast argsort 的输出，而不是让 argsort 改。
- `jt.nn.MaxPool1d` 接受 `return_indices` 但根本不返回索引（5.19 那一类，不是这一类）。
- 审计证据「`pool/core_2d.py:198` 显式 assert 不超过 2^31」的文件名有误：审计当时
  `pool/core_2d.py` 只有 193 行，该断言在 `misc/tensor_ops.py:198`（今天的 214 行）的
  repeat_interleave 里。按内容核实后照修。

+++ python/jittor/pool/core_2d.py
@@ -78,12 +78,12 @@ class Pool(jt.Module):
             else:
                 forward_body += f'''
                 auto out_value = @expand_op(init_{self.op}, @out_type);
-                int out_index = -1;
+                int64 out_index = -1;
                 for (int p = k2; p < k2_; ++p)
                     for (int q = k3; q < k3_; ++q)\x20
                         if (out_value < @in0(i0, i1, p, q)) {{
                             out_value = @in0(i0, i1, p, q);
-                            out_index = p * in0_shape3 + q;
+                            out_index = (int64)p * in0_shape3 + q;
                         }}
                 @out(i0, i1, i2, i3) = out_value;
                 @out1(i0, i1, i2, i3) = out_index;
@@ -105,8 +105,13 @@ class Pool(jt.Module):
                     }}
             '''
             if self.return_indices:
+                # int64 like torch's return_indices, and the encoding
+                # below is computed in int64 too: it is a *flat* offset
+                # into one (n, c) plane, so `p * W + q` leaves int32 as
+                # soon as a plane has 2**31 elements -- which is where
+                # the index would matter most.
                 return_shapes = [[N,C,h,w]] * 2
-                return_dtypes = [x.dtype, 'int32']
+                return_dtypes = [x.dtype, 'int64']
             else:
                 return_shapes = [N,C,h,w]
                 return_dtypes = x.dtype
